feat(message_redraw): Improve boost list message handling
Create a single function, updateSignupReactionMessage(), to replace 
code replicated in 4 different places.

Moved the message redraw functions into a new module to make them easier
to find.
